# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  title: string;
  tagline: string;
  aboutText: string;
  /** @nullable */
  profilePhotoPath: string | null;
  /** @nullable */
  cvPath: string | null;
  updatedAt: string;
}

export interface UpdateProfileBody {
  name?: string;
  title?: string;
  tagline?: string;
  aboutText?: string;
  /** @nullable */
  profilePhotoPath?: string | null;
  /** @nullable */
  cvPath?: string | null;
}

export interface Skill {
  id: number;
  name: string;
  icon: string;
  sortOrder: number;
}

export interface CreateSkillBody {
  name: string;
  icon: string;
  sortOrder?: number;
}

export interface UpdateSkillBody {
  name?: string;
  icon?: string;
  sortOrder?: number;
}

export interface Experience {
  id: number;
  role: string;
  organization: string;
  bullets: string;
  sortOrder: number;
}

export interface CreateExperienceBody {
  role: string;
  organization: string;
  bullets: string;
  sortOrder?: number;
}

export interface UpdateExperienceBody {
  role?: string;
  organization?: string;
  bullets?: string;
  sortOrder?: number;
}

export interface Education {
  id: number;
  institution: string;
  degree: string;
  year: string;
  /** @nullable */
  grade: string | null;
  sortOrder: number;
}

export interface CreateEducationBody {
  institution: string;
  degree: string;
  year: string;
  /** @nullable */
  grade?: string | null;
  sortOrder?: number;
}

export interface UpdateEducationBody {
  institution?: string;
  degree?: string;
  year?: string;
  /** @nullable */
  grade?: string | null;
  sortOrder?: number;
}

export interface Interest {
  id: number;
  name: string;
  icon: string;
  sortOrder: number;
}

export interface CreateInterestBody {
  name: string;
  icon: string;
  sortOrder?: number;
}

export interface UpdateInterestBody {
  name?: string;
  icon?: string;
  sortOrder?: number;
}

export interface Goal {
  id: number;
  content: string;
  sortOrder: number;
}

export interface CreateGoalBody {
  content: string;
  sortOrder?: number;
}

export interface UpdateGoalBody {
  content?: string;
  sortOrder?: number;
}

export interface Contact {
  id: number;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  facebook: string | null;
  /** @nullable */
  instagram: string | null;
  /** @nullable */
  twitter: string | null;
  /** @nullable */
  whatsapp: string | null;
}

export interface UpdateContactBody {
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  whatsapp?: string | null;
}

export interface DashboardStats {
  skillsCount: number;
  experienceCount: number;
  educationCount: number;
  interestsCount: number;
  goalsCount: number;
  hasProfilePhoto: boolean;
  hasCv: boolean;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}